Split exercise controller to satisfy length guard
